+++ src/main/java/ml/treecaptcha/uwuify/core/Configuration.java
@@ -1,0 +1,9 @@
+package ml.treecaptcha.uwuify.core;
+
+public class Configuration {
+    /**
+     * only applicable for use in servers when the uwuifyMessage method is used
+     */
+    public static boolean USE_PREVIEW;
+
+}
