+++ src/main/java/ml/treecaptcha/uwuify/spigot/uwuHandler.java
@@ -1,5 +1,6 @@
-package ml.treecaptcha.uwuify;
+package ml.treecaptcha.uwuify.spigot;
 import io.github.ran.uwu.client.Uwuifier;
+import ml.treecaptcha.uwuify.core.Configuration;
 import org.bukkit.event.Listener;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.block.SignChangeEvent;
@@ -16,7 +17,7 @@ public class uwuHandler implements Listener {
 
     @EventHandler
     public void onPlayerPreview(AsyncPlayerChatPreviewEvent event) {
-        if(!Uwuify.USE_PREVIEW) return;
+        if(!Configuration.USE_PREVIEW) return;
         event.setMessage(Uwuifier.uwuifyMessage(event.getMessage()));
     }
