+++ src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: Uwuify
 version: '${project.version}'
-main: ml.treecaptcha.uwuify.Uwuify
+main: ml.treecaptcha.uwuify.spigot.Uwuify
 api-version: 1.19
 authors: [ TreeCaptcha ]
 description: made with stolen code from Ran-helo/uwu
